Reorganization of templates
