+++ daily_god/apps/frontend/templates/content.html
@@ -13,14 +13,14 @@
     {% for c in content %}
         {% with obj_name=c|class_name %}
             {% if obj_name == 'Post' %}
-                {% include 'components/post_content.html' with post=c %}
+                {% include 'posts/post_content.html' with post=c %}
             {% elif obj_name == 'Prayer' %}
-                {% include 'components/prayer_content.html' with prayer=c %}
+                {% include 'prayers/prayer_content.html' with prayer=c %}
             {% elif obj_name == 'Quote' %}
-                {% include 'components/quote_content.html' with quote=c %}
+                {% include 'quotes/quote_content.html' with quote=c %}
             {% endif %}
         {% endwith %}
     {% endfor %}
 </div>
 
-{% include 'components/new_post.html' %}
+{% include 'posts/new_post.html' %}
