+++ daily_god/apps/templates/posts/new_post.html
