+++ daily_god/apps/templates/quotes/quote_content.html
